moved cache functionality to module
